Followup, 10 items, unneeded CSS
