+++ frontend/pages/SoftwarePage/components/tables/HashCell/_styles.scss
@@ -7,6 +7,7 @@
     display: flex;
     align-items: center;
     margin: -$pad-small 0; // Remove vertical padding but keep clickable area
+    width: 70px; // Consistent to align copy icons
   }
 
   &__copy-overlay {
