+++ frontend/pages/hosts/details/cards/HostSoftwareLibrary/_styles.scss
@@ -28,6 +28,7 @@
       .data-table {
         &__wrapper {
           overflow-x: auto;
+          overflow-y: hidden;
         }
       }
     }
@@ -65,39 +66,8 @@
 
     .data-table-block {
       .data-table__table {
-        thead {
-          .name__header {
-            display: table-cell;
-            width: auto;
-          }
-          .status__header {
-            display: table-cell;
-            width: auto;
-            min-width: 125px;
-          }
-          .Version__header {
-            display: table-cell;
-            width: auto;
-          }
-          .source__header {
-            display: table-cell;
-          }
-          .Vulnerabilities__header {
-            display: table-cell;
-            max-width: px-to-rem(140);
-          }
-          // this is called `id` only due to typing concerns in the table config – it is for the
-          // "Actions" column
-          .Actions__header {
-            width: px-to-rem(72);
-          }
-        }
         tbody {
           tr {
-            .name__cell {
-              display: table-cell;
-              width: auto;
-            }
             .status__cell {
               display: table-cell;
               width: auto;
@@ -106,35 +76,9 @@
                 margin: 0; // Override 10px vertical margin
               }
             }
-            .Version__cell {
-              display: table-cell;
-              width: auto;
-            }
-            .source__cell {
-              display: table-cell;
-            }
-            .Vulnerabilities__cell {
-              display: table-cell;
-            }
 
             .Actions__cell {
-              display: table-cell;
               display: flex;
-            }
-
-            // Very little real estate at 768px
-            @media (max-width: ($break-sm - 1)) {
-              .name__cell {
-                .software-name {
-                  max-width: 190px;
-                }
-              }
-
-              .Version__cell {
-                .text-cell {
-                  max-width: 95px;
-                }
-              }
             }
           }
         }
